Refactor roulette  meme to generator
